FinalProyect/requirements.txt
anyio==3.3.1
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
debugpy==1.4.3
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-heroku==0.3.1
django-six==1.0.4
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
ipykernel==6.4.1
ipython==7.26.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.3
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
kiwisolver==1.3.1
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pyzmq==22.3.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
screen==1.0.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
twitter==1.19.3
unicorn==1.0.3
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.4.7
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
whitenoise==5.3.0
xlwt==1.3.0